Small fixes and resources
Ready for production: Updated the regex for the Brazilian phone mask validation
Ready for production: Added a temporary translate for the "timing reading" stamp for word singularity usage

In development: Created a function to render the image caption inside media and text block. The intention is to use it for the image copyright (but I couldn't render the copyright value yet).